add zkwasm mock, fix input gen, reorg bundle&tests
