+++ bundle-js/common/bundle_local.js
@@ -1,18 +1,7 @@
-var hostMem = ''
+var zkwasmmock = ''
 
-export function setupHostMem(mem){
-    hostMem = mem
-}
-
-function require(a) {
-  if (!a) {
-    console.log("[-] zkwasm require condition is false");
-    throw Error("Abort execution");
-    //TODO: change to graceful kill rather than throw Error?
-  }
-}
-function wasm_input(a) {
-  return hostMem.read_i64();
+export function setupZKWasmMock(mock){
+    zkwasmmock = mock
 }
 
 async function instantiate(module, imports = {}) {
@@ -25,11 +14,11 @@ async function instantiate(module, imports = {}) {
       },
       require(x) {
         // sdk/zkwasm/require1(i32) => i64
-        require(x);
+        zkwasmmock.require(x);
       },
       wasm_input(x) {
         // lib/common/zkwasm/wasm_input(i32) => i64
-        return wasm_input(x) || 0n;
+        return zkwasmmock.wasm_input(x) || 0n;
       },
     }),
   };
