+++ test/lib/test_.js
@@ -1,0 +1,3 @@
+import { test } from "../../build/lib/test.js";
+
+test();
